+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Provide DOT graph of dependencies in the container.
+
 ## [1.7.1] - 2018-09-26
 ### Fixed
 - Make `fxtest.New` ensure that the app was created successfully. Previously,